Count number of triangles in given geometric picture.
Finds the number of triangle in the given geometric shape made with lines. The programe requires the number of lines followed by the lines in point format (For Example: for a triangle ABC the lines are {AB, BC, CA}) and outputs the number of triangles and there notation with respect to there corner points.